@@ -16,19 +16,19 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.3",
-        "symfony/http-kernel": "~2.5",
-        "symfony/dependency-injection": "~2.5",
-        "symfony/http-foundation": "~2.5",
-        "symfony/config": "~2.5",
-        "symfony/form": "~2.5",
-        "symfony/validator": "~2.5",
-        "twig/twig": "~1.11",
+        "php": ">=5.4",
+        "symfony/http-kernel": "^2.5|^3.0",
+        "symfony/dependency-injection": "^2.5|^3.0",
+        "symfony/http-foundation": "^2.5|^3.0",
+        "symfony/config": "^2.5|^3.0",
+        "symfony/form": "^2.5|^3.0",
+        "symfony/validator": "^2.5|^3.0",
+        "twig/twig": "^1.11",
         "fzaninotto/faker": "^1.5"
     },
     "require-dev": {
-        "phpspec/phpspec": "~2.5",
-        "bossa/phpspec2-expect": "~1.0"
+        "phpspec/phpspec": "^2.5",
+        "bossa/phpspec2-expect": "^1.0"
     },
     "config": {
         "bin-dir": "bin"
