@@ -38,6 +38,6 @@ class Dictionary extends Constraint
      */
     public function getRequiredOptions()
     {
-        return array('name');
+        return ['name'];
     }
 }
